+++ NEWS.md
@@ -3,6 +3,7 @@
 **Enhancements**
 
 * [#60](https://github.com/eblondel/geosapi/issues/60) Fix base URL used to retrieve Geoserver version
+* [#61](https://github.com/eblondel/geosapi/issues/61) Support workspace-based methods `getStyles`, `getStyleNames`
 
 
 ## [geosapi 0.6-5](https://cran.r-project.org/package=geosapi) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=geosapi)